Initial Commit
This creates the base for the project with both a web server and a
database for it to connect to.

I chose the Actix Web as the web framework as it seemed to be a popular
and easy to use option.

For the database, I went with sqlite for the same reason.